refactor(scripts): extract shared template-utils helpers
Both `check_template_previews.mjs` and `generate_template_previews.mjs`
inlined the same `listTemplateIds` / `loadTemplateMetadata` /
`isOpenAgreementsOwned` helpers, with their `source_url` predicates having
drifted (only `check_template_previews.mjs` recognized openagreements.org).
Move them to `scripts/lib/template-utils.mjs` so the upcoming
preview-freshness gate (issue #268) can consume one definition.

No behavior change.
